docs(storybook): move stories out of src/
Stories are now on their own root folder, outside src.
