+++ stories/legend.tsx
@@ -1,92 +1,19 @@
 import { storiesOf } from '@storybook/react';
 import React from 'react';
-import { AreaSeries, Axis, Chart, getAxisId, getSpecId, Position, ScaleType, Settings } from '..';
+import {
+  Axis,
+  BarSeries,
+  Chart,
+  getAxisId,
+  getSpecId,
+  Position,
+  ScaleType,
+  Settings,
+} from '../src/';
+import * as TestDatasets from '../src/lib/series/utils/test_dataset';
 
-storiesOf('Area Chart', module)
-  .add('basic', () => {
-    return (
-      <Chart renderer="canvas" className={'story-chart'}>
-        <AreaSeries
-          id={getSpecId('lines')}
-          xScaleType={ScaleType.Linear}
-          yScaleType={ScaleType.Linear}
-          xAccessor="x"
-          yAccessors={['y']}
-          data={[{ x: 0, y: 2 }, { x: 1, y: 7 }, { x: 2, y: 3 }, { x: 3, y: 6 }]}
-          yScaleToDataExtent={false}
-        />
-      </Chart>
-    );
-  })
-  .add('w axis', () => {
-    return (
-      <Chart renderer="canvas" className={'story-chart'}>
-        <Axis
-          id={getAxisId('bottom')}
-          position={Position.Bottom}
-          title={'Bottom axis'}
-          showOverlappingTicks={true}
-        />
-        <Axis
-          id={getAxisId('left2')}
-          title={'Left axis'}
-          position={Position.Left}
-          tickFormat={(d) => Number(d).toFixed(2)}
-        />
-
-        <AreaSeries
-          id={getSpecId('lines')}
-          xScaleType={ScaleType.Linear}
-          yScaleType={ScaleType.Linear}
-          xAccessor="x"
-          yAccessors={['y']}
-          data={[{ x: 0, y: 2 }, { x: 1, y: 7 }, { x: 2, y: 3 }, { x: 3, y: 6 }]}
-          yScaleToDataExtent={false}
-        />
-      </Chart>
-    );
-  })
-  .add('with 4 axes', () => {
-    return (
-      <Chart renderer="canvas" className={'story-chart'}>
-        <Axis
-          id={getAxisId('bottom')}
-          position={Position.Bottom}
-          title={'bottom'}
-          showOverlappingTicks={true}
-        />
-        <Axis
-          id={getAxisId('left')}
-          title={'left'}
-          position={Position.Left}
-          tickFormat={(d) => Number(d).toFixed(2)}
-        />
-        <Axis
-          id={getAxisId('top')}
-          position={Position.Top}
-          title={'top'}
-          showOverlappingTicks={true}
-        />
-        <Axis
-          id={getAxisId('right')}
-          title={'right'}
-          position={Position.Right}
-          tickFormat={(d) => Number(d).toFixed(2)}
-        />
-
-        <AreaSeries
-          id={getSpecId('lines')}
-          xScaleType={ScaleType.Linear}
-          yScaleType={ScaleType.Linear}
-          xAccessor="x"
-          yAccessors={['y']}
-          data={[{ x: 0, y: 2 }, { x: 1, y: 7 }, { x: 2, y: 3 }, { x: 3, y: 6 }]}
-          yScaleToDataExtent={false}
-        />
-      </Chart>
-    );
-  })
-  .add('w axis and legend', () => {
+storiesOf('Legend', module)
+  .add('right', () => {
     return (
       <Chart renderer="canvas" className={'story-chart'}>
         <Settings showLegend={true} legendPosition={Position.Right} />
@@ -103,22 +30,23 @@ storiesOf('Area Chart', module)
           tickFormat={(d) => Number(d).toFixed(2)}
         />
 
-        <AreaSeries
-          id={getSpecId('lines')}
+        <BarSeries
+          id={getSpecId('bars')}
           xScaleType={ScaleType.Linear}
           yScaleType={ScaleType.Linear}
           xAccessor="x"
-          yAccessors={['y']}
-          data={[{ x: 0, y: 2 }, { x: 1, y: 7 }, { x: 2, y: 3 }, { x: 3, y: 6 }]}
+          yAccessors={['y1', 'y2']}
+          splitSeriesAccessors={['g1', 'g2']}
+          data={TestDatasets.BARCHART_2Y2G}
           yScaleToDataExtent={false}
         />
       </Chart>
     );
   })
-  .add('stacked w axis and legend', () => {
+  .add('bottom', () => {
     return (
       <Chart renderer="canvas" className={'story-chart'}>
-        <Settings showLegend={true} legendPosition={Position.Right} />
+        <Settings showLegend={true} legendPosition={Position.Bottom} />
         <Axis
           id={getAxisId('bottom')}
           position={Position.Bottom}
@@ -132,24 +60,74 @@ storiesOf('Area Chart', module)
           tickFormat={(d) => Number(d).toFixed(2)}
         />
 
-        <AreaSeries
-          id={getSpecId('lines')}
+        <BarSeries
+          id={getSpecId('bars')}
           xScaleType={ScaleType.Linear}
           yScaleType={ScaleType.Linear}
           xAccessor="x"
-          yAccessors={['y']}
-          stackAccessors={['x']}
-          splitSeriesAccessors={['g']}
-          data={[
-            { x: 0, y: 2, g: 'a' },
-            { x: 1, y: 7, g: 'a' },
-            { x: 2, y: 3, g: 'a' },
-            { x: 3, y: 6, g: 'a' },
-            { x: 0, y: 4, g: 'b' },
-            { x: 1, y: 5, g: 'b' },
-            { x: 2, y: 8, g: 'b' },
-            { x: 3, y: 2, g: 'b' },
-          ]}
+          yAccessors={['y1', 'y2']}
+          splitSeriesAccessors={['g1', 'g2']}
+          data={TestDatasets.BARCHART_2Y2G}
+          yScaleToDataExtent={false}
+        />
+      </Chart>
+    );
+  })
+  .add('left', () => {
+    return (
+      <Chart renderer="canvas" size={[500, 150]} className={'story-chart'}>
+        <Settings showLegend={true} legendPosition={Position.Left} />
+        <Axis
+          id={getAxisId('bottom')}
+          position={Position.Bottom}
+          title={'Bottom axis'}
+          showOverlappingTicks={true}
+        />
+        <Axis
+          id={getAxisId('left2')}
+          title={'Left axis'}
+          position={Position.Left}
+          tickFormat={(d) => Number(d).toFixed(2)}
+        />
+
+        <BarSeries
+          id={getSpecId('bars')}
+          xScaleType={ScaleType.Linear}
+          yScaleType={ScaleType.Linear}
+          xAccessor="x"
+          yAccessors={['y1', 'y2']}
+          splitSeriesAccessors={['g1', 'g2']}
+          data={TestDatasets.BARCHART_2Y2G}
+          yScaleToDataExtent={false}
+        />
+      </Chart>
+    );
+  })
+  .add('top', () => {
+    return (
+      <Chart renderer="canvas" className={'story-chart'}>
+        <Settings showLegend={true} legendPosition={Position.Top} />
+        <Axis
+          id={getAxisId('bottom')}
+          position={Position.Bottom}
+          title={'Bottom axis'}
+          showOverlappingTicks={true}
+        />
+        <Axis
+          id={getAxisId('left2')}
+          title={'Left axis'}
+          position={Position.Left}
+          tickFormat={(d) => Number(d).toFixed(2)}
+        />
+
+        <BarSeries
+          id={getSpecId('bars')}
+          xScaleType={ScaleType.Linear}
+          yScaleType={ScaleType.Linear}
+          xAccessor="x"
+          yAccessors={['y1', 'y2']}
+          splitSeriesAccessors={['g1', 'g2']}
+          data={TestDatasets.BARCHART_2Y2G}
           yScaleToDataExtent={false}
         />
       </Chart>
